refactor: split out spec processing
